+++ java/src/hype/core/behavior/HTrigger.java
@@ -12,16 +12,18 @@
 package hype.core.behavior;
 
 import hype.core.interfaces.HCallback;
+import hype.core.util.HConstants;
 
 public abstract class HTrigger extends HBehavior {
 	public HCallback _callback;
 	
 	public HTrigger() {
 		register();
+		_callback = HConstants.NOP;
 	}
 	
 	public HTrigger callback(HCallback cb) {
-		_callback = cb;
+		_callback = (cb==null)? HConstants.NOP : cb;
 		return this;
 	}