+++ java/src/hype/extended/behavior/HTimer.java
@@ -92,7 +92,7 @@ public class HTimer extends HTrigger {
 		if(curr-_lastInterval >= _interval) {
 			_lastInterval = curr;
 			
-			if(_callback != null) _callback.run(_cycleCounter);
+			_callback.run(_cycleCounter);
 			
 			if(_numCycles > 0 && ++_cycleCounter >= _numCycles) unregister();
 		}